+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `LaunchImpl.completeLogEmitter()` method, by @HardNorth
+- `Launch.log(SaveLogRQ rq)` method, by @HardNorth
+- `Launch.log(Function<String, SaveLogRQ> logSupplier)` method, by @HardNorth
+### Removed
+- `LaunchLoggingContext` class, by @HardNorth
 
 ## [5.2.31]
 ### Added